Fixed conflicting declarations in cpu.h

+++ src/cpu.cpp
@@ -66,22 +66,22 @@ void GBA::CPU::reset() {
 
 uint32_t& GBA::CPU::SP(GBA::CPU::Mode mode) {
     if (mode == GBA::CPU::Mode::User || mode == GBA::CPU::Mode::System) {
-        return registers[GBA::CPU::RegisterIndex::SP];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP)];
     }
     else if (mode == GBA::CPU::Mode::FastInterrupt) {
-        return registers[GBA::CPU::RegisterIndex::SP_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_FIQ)];
     }
     else if (mode == GBA::CPU::Mode::Supervisor) {
-        return registers[GBA::CPU::RegisterIndex::SP_SVC];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_SVC)];
     }
     else if (mode == GBA::CPU::Mode::Abort) {
-        return registers[GBA::CPU::RegisterIndex::SP_ABT];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_ABT)];
     }
     else if (mode == GBA::CPU::Mode::Interrupt) {
-        return registers[GBA::CPU::RegisterIndex::SP_IRQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_IRQ)];
     }
     else if (mode == GBA::CPU::Mode::Undefined) {
-        return registers[GBA::CPU::RegisterIndex::SP_UND];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_UND)];
     }
     // TODO: invalid mode error handling
     throw;
@@ -89,22 +89,22 @@ uint32_t& GBA::CPU::SP(GBA::CPU::Mode mode) {
 
 const uint32_t& GBA::CPU::SP(GBA::CPU::Mode mode) const {
     if (mode == GBA::CPU::Mode::User || mode == GBA::CPU::Mode::System) {
-        return registers[GBA::CPU::RegisterIndex::SP];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP)];
     }
     else if (mode == GBA::CPU::Mode::FastInterrupt) {
-        return registers[GBA::CPU::RegisterIndex::SP_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_FIQ)];
     }
     else if (mode == GBA::CPU::Mode::Supervisor) {
-        return registers[GBA::CPU::RegisterIndex::SP_SVC];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_SVC)];
     }
     else if (mode == GBA::CPU::Mode::Abort) {
-        return registers[GBA::CPU::RegisterIndex::SP_ABT];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_ABT)];
     }
     else if (mode == GBA::CPU::Mode::Interrupt) {
-        return registers[GBA::CPU::RegisterIndex::SP_IRQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_IRQ)];
     }
     else if (mode == GBA::CPU::Mode::Undefined) {
-        return registers[GBA::CPU::RegisterIndex::SP_UND];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::SP_UND)];
     }
     // TODO: invalid mode error handling
     throw;
@@ -112,22 +112,22 @@ const uint32_t& GBA::CPU::SP(GBA::CPU::Mode mode) const {
 
 uint32_t& GBA::CPU::LR(GBA::CPU::Mode mode) {
     if (mode == GBA::CPU::Mode::User || mode == GBA::CPU::Mode::System) {
-        return registers[GBA::CPU::RegisterIndex::LR];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR)];
     }
     else if (mode == GBA::CPU::Mode::FastInterrupt) {
-        return registers[GBA::CPU::RegisterIndex::LR_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_FIQ)];
     }
     else if (mode == GBA::CPU::Mode::Supervisor) {
-        return registers[GBA::CPU::RegisterIndex::LR_SVC];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_SVC)];
     }
     else if (mode == GBA::CPU::Mode::Abort) {
-        return registers[GBA::CPU::RegisterIndex::LR_ABT];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_ABT)];
     }
     else if (mode == GBA::CPU::Mode::Interrupt) {
-        return registers[GBA::CPU::RegisterIndex::LR_IRQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_IRQ)];
     }
     else if (mode == GBA::CPU::Mode::Undefined) {
-        return registers[GBA::CPU::RegisterIndex::LR_UND];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_UND)];
     }
     // TODO: invalid mode error handling
     throw;
@@ -135,33 +135,33 @@ uint32_t& GBA::CPU::LR(GBA::CPU::Mode mode) {
 
 const uint32_t& GBA::CPU::LR(GBA::CPU::Mode mode) const {
     if (mode == GBA::CPU::Mode::User || mode == GBA::CPU::Mode::System) {
-        return registers[GBA::CPU::RegisterIndex::LR];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR)];
     }
     else if (mode == GBA::CPU::Mode::FastInterrupt) {
-        return registers[GBA::CPU::RegisterIndex::LR_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_FIQ)];
     }
     else if (mode == GBA::CPU::Mode::Supervisor) {
-        return registers[GBA::CPU::RegisterIndex::LR_SVC];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_SVC)];
     }
     else if (mode == GBA::CPU::Mode::Abort) {
-        return registers[GBA::CPU::RegisterIndex::LR_ABT];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_ABT)];
     }
     else if (mode == GBA::CPU::Mode::Interrupt) {
-        return registers[GBA::CPU::RegisterIndex::LR_IRQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_IRQ)];
     }
     else if (mode == GBA::CPU::Mode::Undefined) {
-        return registers[GBA::CPU::RegisterIndex::LR_UND];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::LR_UND)];
     }
     // TODO: invalid mode error handling
     throw;
 }
 
 uint32_t& GBA::CPU::PC() {
-    return registers[GBA::CPU::RegisterIndex::PC];
+    return registers[static_cast<int>(GBA::CPU::RegisterIndex::PC)];
 }
 
 const uint32_t& GBA::CPU::PC() const {
-    return registers[GBA::CPU::RegisterIndex::PC];
+    return registers[static_cast<int>(GBA::CPU::RegisterIndex::PC)];
 }
 
 uint32_t& GBA::CPU::R(uint32_t index) {
@@ -193,19 +193,19 @@ uint32_t& GBA::CPU::R_FIQ(uint32_t index) {
     case 7:
         return registers[index];
     case 8:
-        return registers[GBA::CPU::RegisterIndex::R8_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R8_FIQ)];
     case 9:
-        return registers[GBA::CPU::RegisterIndex::R9_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R9_FIQ)];
     case 10:
-        return registers[GBA::CPU::RegisterIndex::R10_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R10_FIQ)];
     case 11:
-        return registers[GBA::CPU::RegisterIndex::R11_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R11_FIQ)];
     case 12:
-        return registers[GBA::CPU::RegisterIndex::R12_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R12_FIQ)];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_FIQ)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_FIQ)];
     default:
         // TODO: error handling
         throw;
@@ -225,19 +225,19 @@ const uint32_t& GBA::CPU::R_FIQ(uint32_t index) const {
     case 7:
         return registers[index];
     case 8:
-        return registers[GBA::CPU::RegisterIndex::R8_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R8_FIQ)];
     case 9:
-        return registers[GBA::CPU::RegisterIndex::R9_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R9_FIQ)];
     case 10:
-        return registers[GBA::CPU::RegisterIndex::R10_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R10_FIQ)];
     case 11:
-        return registers[GBA::CPU::RegisterIndex::R11_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R11_FIQ)];
     case 12:
-        return registers[GBA::CPU::RegisterIndex::R12_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R12_FIQ)];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_FIQ)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_FIQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_FIQ)];
     default:
         // TODO: error handling
         throw;
@@ -262,9 +262,9 @@ uint32_t& GBA::CPU::R_SVC(uint32_t index) {
     case 12:
         return registers[index];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_SVC];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_SVC)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_SVC];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_SVC)];
     default:
         // TODO: error handling
         throw;
@@ -289,9 +289,9 @@ const uint32_t& GBA::CPU::R_SVC(uint32_t index) const {
     case 12:
         return registers[index];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_SVC];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_SVC)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_SVC];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_SVC)];
     default:
         // TODO: error handling
         throw;
@@ -316,9 +316,9 @@ uint32_t& GBA::CPU::R_ABT(uint32_t index) {
     case 12:
         return registers[index];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_ABT];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_ABT)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_ABT];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_ABT)];
     default:
         // TODO: error handling
         throw;
@@ -343,9 +343,9 @@ const uint32_t& GBA::CPU::R_ABT(uint32_t index) const {
     case 12:
         return registers[index];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_ABT];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_ABT)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_ABT];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_ABT)];
     default:
         // TODO: error handling
         throw;
@@ -370,9 +370,9 @@ uint32_t& GBA::CPU::R_IRQ(uint32_t index) {
     case 12:
         return registers[index];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_IRQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_IRQ)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_IRQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_IRQ)];
     default:
         // TODO: error handling
         throw;
@@ -397,9 +397,9 @@ const uint32_t& GBA::CPU::R_IRQ(uint32_t index) const {
     case 12:
         return registers[index];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_IRQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_IRQ)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_IRQ];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_IRQ)];
     default:
         // TODO: error handling
         throw;
@@ -424,9 +424,9 @@ uint32_t& GBA::CPU::R_UND(uint32_t index) {
     case 12:
         return registers[index];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_UND];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_UND)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_UND];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_UND)];
     default:
         // TODO: error handling
         throw;
@@ -451,9 +451,9 @@ const uint32_t& GBA::CPU::R_UND(uint32_t index) const {
     case 12:
         return registers[index];
     case 13:
-        return registers[GBA::CPU::RegisterIndex::R13_UND];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R13_UND)];
     case 14:
-        return registers[GBA::CPU::RegisterIndex::R14_UND];
+        return registers[static_cast<int>(GBA::CPU::RegisterIndex::R14_UND)];
     default:
         // TODO: error handling
         throw;
